requirements.txt
torch>=2.0
numpy>=1.21
rich>=12.0
pyyaml>=6.0
pybind11>=2.6
onnx>=1.14
onnxruntime>=1.15
